Post delete feature implemented

+++ Server/src/controllers/post/postController.js
@@ -89,8 +89,21 @@ const deletePost = asyncHandler(async (req, res) => {
   const { postId } = req.params;
 
   try {
-    const deletedPost = await Post.findByIdAndDelete(postId);
-    res.status(200).json({ message: "Post Deleted!" });
+    const post = await Post.findById(postId);
+
+    if (!post) {
+      return res.status(404).json({ message: "Post not found" });
+    }
+
+    // Check if the current user's _id matches the post's userId
+    if (req.user._id.toString() === post.userId.toString()) {
+      const deletedPost = await Post.findByIdAndDelete(postId);
+      return res.status(200).json({ message: "Post Deleted!" });
+    } else {
+      return res
+        .status(403)
+        .json({ message: "You are not authorized to delete this post." });
+    }
   } catch (error) {
     throw new Error(error);
   }
